backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
openai==1.55.3
httpx==0.27.2
moviepy==2.1.2
pillow==10.4.0
decorator==5.1.1
imageio==2.36.1
imageio_ffmpeg==0.5.1
proglog==0.1.10
setuptools==75.8.0
moviepy==2.1.2
psycopg2-binary==2.9.6
ffmpeg-python==0.2.0
gTTS==2.5.3
numpy==2.1.1
pydantic==2.9.2
python-dotenv==1.0.1
cloudinary==1.41.0
watchdog==6.0.0
